requirements.txt
# AI Presentation Agent - Dependencies
# Python 3.11+

openai>=1.12.0
httpx>=0.27.0
pydantic>=2.6.0
python-pptx>=0.6.23
Pillow>=10.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
